requirements.txt
python-dotenv
requests
spond
